final task 2nd attempt

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,11 @@ int main(void)
 	long int long_int;
 	long long int long_long_int;
 	float float_type;
-	printf("size of a char: %zu byte(s)\n", sizeof(char_type));
-	printf("size of an int: %zu byte(s)\n", sizeof(int_type));
-	printf("size of a long int: %zu byte(s)\n", sizeof(long_int));
-	printf("size of a long long int: %zu byte(s)\n", sizeof(long_long_int));
-	printf("size of a float: %zu byte(s)\n", sizeof(float_type));
+       
+       	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(char_type));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(int_type));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long_int));
+	printf("size of a long long int: %lu byte(s)\n",(unsigned long) sizeof(long_long_int));
+	printf("size of a float: %lu byte(s)\n",(unsigned long) sizeof(float_type));
 	return (0);
 }
